Globals.py
import os
from dotenv import load_dotenv
#strings

load_dotenv()

DAVINCI_TOKEN = os.environ.get('DAVINCI_TOKEN')

SERVER_ID = os.environ.get('SERVER_ID')

SALAI_TOKEN = os.environ.get('SALAI_TOKEN')

CHANNEL_ID = os.environ.get('CHANNEL_ID')


#boolean
USE_MESSAGED_CHANNEL = False

#don't edit the following variable
MID_JOURNEY_ID = "936929561302675456"  #midjourney bot id
targetID       = ""
targetHash     = ""